innit

+++ cat/message/event.go
@@ -4,7 +4,7 @@ type Event struct {
 	baseMessage
 }
 
-func NewEvent(t, name, status, data string, timestampInMillis int) *Event {
+func NewEvent(t, name, status, data string, timestampInMillis int64) *Event {
 	return &Event{
 		baseMessage: newBaseMessage(t, name, status, data, timestampInMillis),
 	}
